refactor: extract LocationSheet into manager architecture
Break down 755-line LocationSheet into focused managers:
- LocationSheet: 149 lines (orchestration only)
- ItemManager: 433 lines (all item/inventory logic)
- DataManager: 163 lines (data preparation)
- UIManager: 123 lines (rendering/UI manipulation)

Benefits:
- 80% reduction in main sheet complexity
- Testable separation of concerns
- Clean extension points for future features
- Beautiful maintainable architecture
